+++ gems/aws-partitions/lib/aws-partitions/partition_list.rb
@@ -90,6 +90,7 @@ module Aws
           else
             regions[region_name] = Region.new(
               name: region_name,
+
               description: region['description'],
               partition_name: partition_name,
               services: Set.new