+++ gems/aws-sdk-core/lib/aws-sdk-core/endpoints/matchers.rb
@@ -94,14 +94,15 @@ module Aws
 
       # aws.partition(value: string) Option<Partition>
       def self.aws_partition(value)
-        partition =
-          Aws::Partitions.find { |p| p.region?(value) } ||
-          Aws::Partitions.find { |p| value.match(p.region_regex) } ||
-          Aws::Partitions.find { |p| p.name == 'aws' }
-
-        return nil unless partition
-
-        partition.metadata
+        Aws::Partitions::Metadata.aws_partition(value)
+        # partition =
+        #   Aws::Partitions.find { |p| p.region?(value) } ||
+        #   Aws::Partitions.find { |p| value.match(p.region_regex) } ||
+        #   Aws::Partitions.find { |p| p.name == 'aws' }
+        #
+        # return nil unless partition
+        #
+        # partition.metadata
       end
 
       # aws.parseArn(value: string) Option<ARN>